test: Add new lint tests for module changes, deprecations, todos, and version handling
- Introduced separate test files for module changes, deprecations, todos, and version functionalities.
- Implemented basic structure and placeholder tests for each new test file to ensure future functionality can be added.
- Removed the old test_lint.py file to streamline the test organization.
