+++ tests/modules/lint/test_module_changes.py
@@ -1,0 +1,66 @@
+import json
+import shutil
+from pathlib import Path
+from typing import Union
+
+import pytest
+import yaml
+from git.repo import Repo
+
+import nf_core.modules.lint
+from nf_core.modules.lint.module_changes import module_changes
+from nf_core.utils import set_wd
+
+from ...test_modules import TestModules
+from ...utils import GITLAB_NFTEST_BRANCH, GITLAB_URL
+
+
+# A skeleton object with the passed/warned/failed list attrs
+# Use this in place of a ModuleLint object to test behaviour of
+# linting methods which don't need the full setup
+class MockModuleLint:
+    def __init__(self):
+        self.passed = []
+        self.warned = []
+        self.failed = []
+        self.main_nf = "main_nf"
+
+
+class TestModuleChanges(TestModules):
+    """Test module_changes.py functionality"""
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_changes_unchanged(self):
+        """Test module changes when module is unchanged"""
+        # Test the functionality of module_changes.py when module is unchanged
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_changes_modified(self):
+        """Test module changes when module is modified"""
+        # Test the functionality of module_changes.py when module is modified
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_changes_patched(self):
+        """Test module changes when module is patched"""
+        # Test when module has patches applied
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_changes_main_nf_modified(self):
+        """Test module changes when main.nf is modified"""
+        # Test when main.nf file is modified
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_changes_meta_yml_modified(self):
+        """Test module changes when meta.yml is modified"""
+        # Test when meta.yml file is modified
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_changes_patch_apply_fail(self):
+        """Test module changes when patch application fails"""
+        # Test when patch cannot be applied in reverse
+        pass
