+++ tests/modules/lint/test_module_deprecations.py
@@ -1,0 +1,54 @@
+import json
+import shutil
+from pathlib import Path
+from typing import Union
+
+import pytest
+import yaml
+from git.repo import Repo
+
+import nf_core.modules.lint
+from nf_core.modules.lint.module_deprecations import module_deprecations
+from nf_core.utils import set_wd
+
+from ...test_modules import TestModules
+from ...utils import GITLAB_NFTEST_BRANCH, GITLAB_URL
+
+
+# A skeleton object with the passed/warned/failed list attrs
+# Use this in place of a ModuleLint object to test behaviour of
+# linting methods which don't need the full setup
+class MockModuleLint:
+    def __init__(self):
+        self.passed = []
+        self.warned = []
+        self.failed = []
+        self.main_nf = "main_nf"
+
+
+class TestModuleDeprecations(TestModules):
+    """Test module_deprecations.py functionality"""
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_deprecations_none(self):
+        """Test module deprecations when no deprecations exist"""
+        # Test the functionality of module_deprecations.py when no deprecated files exist
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_deprecations_found(self):
+        """Test module deprecations when deprecations are found"""
+        # Test the functionality of module_deprecations.py when deprecated files are found
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_deprecations_functions_nf(self):
+        """Test module deprecations when functions.nf exists"""
+        # Test when deprecated functions.nf file is found
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_deprecations_no_functions_nf(self):
+        """Test module deprecations when no functions.nf exists"""
+        # Test when no deprecated files are found
+        pass
