+++ tests/modules/lint/test_module_todos.py
@@ -1,0 +1,54 @@
+import json
+import shutil
+from pathlib import Path
+from typing import Union
+
+import pytest
+import yaml
+from git.repo import Repo
+
+import nf_core.modules.lint
+from nf_core.modules.lint.module_todos import module_todos
+from nf_core.utils import set_wd
+
+from ...test_modules import TestModules
+from ...utils import GITLAB_NFTEST_BRANCH, GITLAB_URL
+
+
+# A skeleton object with the passed/warned/failed list attrs
+# Use this in place of a ModuleLint object to test behaviour of
+# linting methods which don't need the full setup
+class MockModuleLint:
+    def __init__(self):
+        self.passed = []
+        self.warned = []
+        self.failed = []
+        self.main_nf = "main_nf"
+
+
+class TestModuleTodos(TestModules):
+    """Test module_todos.py functionality"""
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_todos_none(self):
+        """Test module todos when no TODOs exist"""
+        # Test the functionality of module_todos.py when no TODO statements are found
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_todos_found(self):
+        """Test module todos when TODOs are found"""
+        # Test the functionality of module_todos.py when TODO statements are found
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_todos_markdown(self):
+        """Test module todos when markdown TODOs exist"""
+        # Test finding TODO statements in markdown files
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_todos_groovy(self):
+        """Test module todos when groovy TODOs exist"""
+        # Test finding TODO statements in Nextflow/Groovy files
+        pass
