+++ tests/modules/lint/test_module_version.py
@@ -1,0 +1,66 @@
+import json
+import shutil
+from pathlib import Path
+from typing import Union
+
+import pytest
+import yaml
+from git.repo import Repo
+
+import nf_core.modules.lint
+from nf_core.modules.lint.module_version import module_version
+from nf_core.utils import set_wd
+
+from ...test_modules import TestModules
+from ...utils import GITLAB_NFTEST_BRANCH, GITLAB_URL
+
+
+# A skeleton object with the passed/warned/failed list attrs
+# Use this in place of a ModuleLint object to test behaviour of
+# linting methods which don't need the full setup
+class MockModuleLint:
+    def __init__(self):
+        self.passed = []
+        self.warned = []
+        self.failed = []
+        self.main_nf = "main_nf"
+
+
+class TestModuleVersion(TestModules):
+    """Test module_version.py functionality"""
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_version_valid(self):
+        """Test module version when version is valid"""
+        # Test the functionality of module_version.py when version is valid
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_version_invalid(self):
+        """Test module version when version is invalid"""
+        # Test the functionality of module_version.py when version is invalid
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_version_up_to_date(self):
+        """Test module version when module is up to date"""
+        # Test when module is at the latest version
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_version_outdated(self):
+        """Test module version when module is outdated"""
+        # Test when module has newer version available
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_version_no_git_sha(self):
+        """Test module version when no git_sha in modules.json"""
+        # Test when modules.json is missing git_sha entry
+        pass
+
+    @pytest.mark.skip(reason="Test implementation pending")
+    def test_module_version_git_log_fail(self):
+        """Test module version when git log fetch fails"""
+        # Test when fetching git log fails
+        pass
